ssl v3 integration done to show balances on ui + ui fixes
